// pwa-app/src/types/types.ts
export interface GeofenceSettings {
  IsGeoFencingEnabled: boolean; // GeoLocation feature ON/OFF Client wise    // server returns string so converting it to boolean to use in app flow
  IsFetchOneTimeLatLogEnabled: boolean; // Capture Shop Location ON/OFF             // server returns string so converting it to boolean to use in app flow
  IsLocationRestriction: boolean; // Fence Restricition for Shop ON/OFF       // server returns string so converting it to boolean to use in app flow
  GeofenceRadius_Shop?: string; // Phase 4 implementation
  GeofenceRadiusLatDelat?: string; // Phase 4 implementation
  GeofenceRadiusLongDelat?: string; // Phase 4 implementation
  GeoFenceDistanceFilterMtr?: string; // To call library to fetch onlocation after desired mtr
  IsLiveLocationTracking: boolean; // if true then enable live location tracking
  LiveLocationStartTime: string;
  LiveLocationEndTime: string;
  LoggedInUserId: string;
  IsShopEnteredNotificationEnabled: boolean; // to display notification on entering geofenced area.
  LastSyncLiveLocationApiTimeStamp: number; //
  LastSyncThresholdTimeApiCall: number; // in min, to call api after the last location api call has happen in interval.
  TypeOfLiveLocationTracking: TypeOfLiveLocationTracking;
  ActivityTriggered: ActivityTirggeredTracking;
}

export interface TypeOfLiveLocationTracking {
  IsAttandanceBasedTracking: boolean;
  IsAppOpenedActivityBasedTracking: boolean;
}

export interface ActivityTirggeredTracking {
  IsAttandanceTriggeredForTheDay: boolean;
  AppOpenedActivityTriggeredForTheDay: string | undefined;
}